+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to Instanced Audio Adjuster will be documented in this file.
 ## [12.0.1-3]
 
 - Major Refactor of the codebase, fully backwards compatible
+- Update options menu to be more user friendly
 
 
 ## [12.0.1-2] - 2026-02-03